requirements.txt
mysql-connector-python-rf
Flask
Flask-Session
requests
PIL
